Implement CacaoCreation in main contract, StandardToken methods now verify valid ammounts, creation and destruction can now be run only if not frozen

+++ contracts/StandardToken.sol
@@ -120,7 +120,7 @@ contract StandardToken is ERC20 {
     /// To increment allowed value is better to use this function to avoid 2 calls (and wait until the first transaction is mined)
     /// @param _spender The address which will spend the funds.
     /// @param _addedValue The amount of tokens to increase the allowance by.
-    function increaseApproval(address _spender, uint _addedValue) public returns (bool) {
+    function increaseApproval(address _spender, uint256 _addedValue) public returns (bool) {
         allowed[msg.sender][_spender] = allowed[msg.sender][_spender].add(_addedValue);
         emit Approval(msg.sender, _spender, allowed[msg.sender][_spender]);
         return true;
@@ -131,8 +131,8 @@ contract StandardToken is ERC20 {
    /// To decrement allowed value is better to use this function to avoid 2 calls (and wait until the first transaction is mined)
    /// @param _spender The address which will spend the funds.
    /// @param _subtractedValue The amount of tokens to decrease the allowance by.
-    function decreaseApproval(address _spender, uint _subtractedValue) public returns (bool) {
-        uint oldValue = allowed[msg.sender][_spender];
+    function decreaseApproval(address _spender, uint256 _subtractedValue) public returns (bool) {
+        uint256 oldValue = allowed[msg.sender][_spender];
         if (_subtractedValue > oldValue) {
             allowed[msg.sender][_spender] = 0;
         } else {
